feat!: change listing stragety
